Initial commit of the DAP internal fork of geometry3Sharp to public Github.

+++ approximation/GaussPointsFit3.cs
@@ -28,8 +28,8 @@ namespace g3
             Box.Center *= invNumPoints;
 
             // Compute the covariance matrix of the points.
-            double sumXX = (double)0, sumXY = (double)0, sumXZ = (double)0;
-            double sumYY = (double)0, sumYZ = (double)0, sumZZ = (double)0;
+            double sumXX = 0, sumXY = 0, sumXZ = 0;
+            double sumYY = 0, sumYZ = 0, sumZZ = 0;
             foreach (Vector3d p in points) { 
                 Vector3d diff = p - Box.Center;
                 sumXX += diff[0] * diff[0];
@@ -66,8 +66,8 @@ namespace g3
             Box.Center *= invWeightDivide;
 
             // Compute the covariance matrix of the points.
-            double sumXX = (double)0, sumXY = (double)0, sumXZ = (double)0;
-            double sumYY = (double)0, sumYZ = (double)0, sumZZ = (double)0;
+            double sumXX = 0, sumXY = 0, sumXZ = 0;
+            double sumYY = 0, sumYZ = 0, sumZZ = 0;
             weights_itr = weights.GetEnumerator();
             foreach (Vector3d p in points) {
                 weights_itr.MoveNext();
