+++ core/TagSet.cs
@@ -60,7 +60,7 @@ namespace g3
             type = (byte)(combined >> 24);
             value = combined & 0xFFFFFF;
         }
-        public int intValue { get { return ((int)type) << 24 | value; } }
+        public int intValue { get { return type << 24 | value; } }
     }
