+++ intersection/IntrSegment3Box3.cs
@@ -197,7 +197,7 @@ namespace g3
 			// plane.  Otherwise 'false' is returned in which case the line segment
 			// is entirely clipped.
 
-			if (denom > (double)0)
+			if (denom > 0)
 			{
 				if (numer > denom*t1)
 				{
@@ -209,7 +209,7 @@ namespace g3
 				}
 				return true;
 			}
-			else if (denom < (double)0)
+			else if (denom < 0)
 			{
 				if (numer > denom*t0)
 				{
@@ -223,7 +223,7 @@ namespace g3
 			}
 			else
 			{
-				return numer <= (double)0;
+				return numer <= 0;
 			}
 		}
